# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parking Lot LObby",
    "description": "PLL Api Documentation",
    "termsOfService": "",
    "contact": {
      "name": "matrix marketer",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1"
  },
...
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
    },
    "schemas": {
...
      "TwilioSms":{
        "type":"object",
        "properties": {
          "Body": {
            "type": "string",
            "required":true
          },
          "To": {
            "type": "string",
            "description":"privde clinic(provider) contact no.",
            "required":true
          },
          "From": {
            "type": "string",
            "description":"privde patient contact no.",
            "required":true
          }
        },
        "required": [
          "Body",
          "To",
          "From"
        ]
      },
      "TwilioSend":{
        "type":"object",
        "properties": {
          "patientId": {
            "type": "string",
            "required":true
          },
          "message": {
            "type": "string",
            "required":true
          }
        },
        "required": [
          "patientId",
          "message"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/twilio/sms": {
      "post": {
        "summary": "For test webhook response",
        "description": "This Api will use by twilio for webhook",
        "tags": [
          "Twillio List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwilioSms"
              }
            }
          }
        }
      }
    },
    "/twilio/send": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Send Message to patient",
        "description": "This for send a message to Patient on behalf of clinic",
        "tags": [
          "Twillio List"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/TwilioSend"
              }
            }
          }
        }
      }
    },
...
  }
}